+++ effective-shell-playground/scripts/nobang.sh
@@ -1,3 +1,0 @@
-# nobang: This script shows an anti-pattern - not using a shebang in a shell
-# script. It shows the process tree to for the shell that runs the script:
-pstree -a -s $$